refactor: 플레이어 수를 검증로직 책임 분리
- GameManager를 도입하여 기존 InputView에 있던 Player검증로직을 분리하였습니다.
